requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
python-dotenv>=1.0
psycopg[binary]>=3.1
gunicorn>=21.2
whitenoise>=6.6
